+++ lib/rubocop/cop/style/space_around_keyword.rb
@@ -182,11 +182,7 @@ module RuboCop
           return false unless ancestor
           return true if ancestor.and_type? || ancestor.or_type?
           return false unless ancestor.send_type?
-          operator?(ancestor.loc.selector)
-        end
-
-        def operator?(range)
-          range.source !~ /^\[|\w/
+          operator?(ancestor.loc.selector.source.to_sym)
         end
 
         def autocorrect(range)